Move to a more modular approach for Javascript
We know build our JS with grunt. We lint with eslint, typecheck with flow, and build with babel/browserify.

This diff removes clock.js from the gameboard, to be added back later. Some minor code cleanup was done based on flow/eslint, but this diff is more getting the build system setup so we can do incremental improvements later.